Removed the call to iop-strip for make-release and added the use of -mno-explicit-relocs.
The call to iop-strip resulted in broken IRX files. It's probably just the wrong way of stripping an IRX, but I feel that Binutils needs some work too.

As for -mno-explicit-relocs, it's to work around the fact that GCC is now known
to output multiple LO relocations after one HI relocation. While this is legal
under the standard MIPS ABI, the PlayStation 2 IOP kernel does not support that.

With this option, GCC would output normal assembly instructions instead and iop-as would deal with the relocations on its own.
As a side effect (it probably always was), instructions like "la $XX, <label>" would get generated instead.
That seems to always result in a LO relocation following a HI relocation.